requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
langchain-core==0.2.38
tavily-python==0.5.0
python-dotenv
colorama